+++ frontend/src/pages/Home.jsx
@@ -2,14 +2,13 @@ import { useEffect, useState } from "react";
 import categories from "../assets/categories";
 import products from "../assets/products"
 import { CiHeart } from "react-icons/ci";
-import { FaLaptopCode } from "react-icons/fa";
-import { CiDesktopMouse1 } from "react-icons/ci";
+import BestSelling from "../components/BestSelling";
 
 const Home = () => {
 
     const [category, setCategory] = useState([])
     const [productsList, setProductsList] = useState(products)
-    const [likedProducts, setLikedProducts] = useState([])
+    
 
     const toggleLike = (id) => {
         setProductList((prevProducts) =>
@@ -27,13 +26,7 @@ const Home = () => {
     }, [])
 
 
-    useEffect(() => {
-        const fetchLiked = () => {
-            const likedProducts = products.filter((product) => product.liked === true)
-            setLikedProducts(likedProducts)
-        }
-        fetchLiked()
-    }, [])
+    
     return (
         <div className="p-10">
             <div className="grid grid-cols-4 gap-4">
@@ -91,7 +84,7 @@ const Home = () => {
                             <p className="text-3xl text-start"> Browse By Category </p>
                             <div className="grid grid-cols-4 gap-6 mt-10">
                                 {categories.map((cat) => (
-                                    <div className="border border-gray-400 w-30 h-full rounded flex flex-col items-center p-6 cursor-pointer" key={cat.id}>
+                                    <div className="border border-gray-400 w-30 rounded flex flex-col items-center p-6 cursor-pointer" key={cat.id}>
                                         <img src={cat.image} alt="" />
                                         <p> {cat.name} </p>
                                     </div>
@@ -110,32 +103,12 @@ const Home = () => {
                     </div>
 
                     <div className="">
-                        <div className="flex justify-between items-center">
-                            <p className="text-3xl"> Best Selling Products</p>
-                            <button className="bg-[#DB4444] text-white px-5 py-2"> View All </button>
-                        </div>
-                        <div className="mt-10">
-                            <div className="grid grid-cols-4 gap-7">
-                                {
-                                    likedProducts.map((product) => (
-                                        <div key={product.id} className="h-80 bg-green-500 w-full p-5 flex flex-col gap-6">
-                                            <div key={product.id} className="h-80 bg-green-500 w-full p-5">
-                                                <div className="relative h-40">
-                                                    <CiHeart
-                                                        onClick={() => toggleLike(product.id)}
-                                                        className={`absolute right-0 top-0 text-2xl cursor-pointer ${product.liked ? "text-red-600" : "text-white"
-                                                            }`}
-                                                    />
-                                                    <img src={product.image[0].frontUrl} alt="" />
-                                                </div>
-                                                <small>{product.name}</small>
-                                                <p>Ksh{product.price}</p>
-                                            </div>
-                                        </div>
-                                    ))
-                                }
-                            </div>
-                        </div>
+                        <BestSelling productsList={productsList} toggleLike={toggleLike}/>                      
+                    </div>
+
+
+                    <div className="mt-20 flex justify-center">
+                        <img src="/images/Frame 600.png" alt="" />
                     </div>
                 </div>
             </div>
